requirements.txt
fastapi
uvicorn[standard]
ccxt
httpx
python-dotenv
gspread
google-auth
